Move save to own file
